api/requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.1
cloudinary==1.41.0
colorama==0.4.6
distlib==0.3.9
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
filelock==3.16.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
matplotlip==0.2
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.4
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.0
